Refactored Condo Form into a reusable component
